Extract fields from user sale profile
