+++ src/pages/users/components/form/index.tsx
@@ -3,6 +3,7 @@ import { CollectorProfileForm } from "./CollectorProfileForm"
 import { UserAdminForm } from "./UserAdminForm"
 import { UserFairActionsForm } from "./UserFairActionsForm"
 import { UserInfoForm } from "./UserInfoForm"
+import { UserSaleProfileForm } from "./UserSaleProfileForm"
 import { UserInterestsForm } from "./UserInterestsForm"
 import { UserTagsForm } from "./UserTagsForm"
 
@@ -12,6 +13,7 @@ export const UserForm: React.FC = () => {
       <GridColumns>
         <Column span={4}>
           <UserInfoForm />
+          <UserSaleProfileForm />
         </Column>
         <Column span={4} px={4}>
           <CollectorProfileForm />
